requirements.txt
requests~=2.32.3
pytest~=8.3.3
factory_boy~=3.3.1
Faker~=29.0.0
